Update API skeleton
